backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-multipart==0.0.20
reportlab==4.2.2
rapidocr-onnxruntime==1.4.4
pillow==12.1.1
